Radar (complete) and slug rifle WIP

+++ source/src/protos.h
@@ -630,7 +630,7 @@ extern void clearbounceents();
 extern void renderbounceents();
 extern void addgib(playerent *d);
 extern playerent *playerincrosshairhit(int &hitzone);
-extern int magsize(int gun);
+extern ushort magsize(int gun);
 extern void checkweaponswitch();
 extern void setscope(bool activate);
 extern int intersect(playerent *d, const vec &from, const vec &to, vec *end = NULL);
